+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Data/Models/SignificantChange/EqualitiesImpact.cs
@@ -1,0 +1,8 @@
+namespace Dfe.PrepareConversions.Data.Models.SignificantChange;
+
+public enum EqualitiesImpact
+{
+   None,
+   PotentialImpacts,
+   ImpactsIdentified
+}