+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Data/Models/SignificantChange/SetSignificantChangeEqualitiesImpactAssessmentCommand.cs
@@ -1,0 +1,6 @@
+namespace Dfe.PrepareConversions.Data.Models.SignificantChange;
+
+public record SetSignificantChangeEqualitiesImpactAssessmentCommand(
+   bool? EqualitiesImpactAssessmentCompleted,
+   EqualitiesImpact? EqualitiesImpactIdentified,
+   string EqualitiesImpactIdentifiedMitigation);