+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Data/Models/SignificantChange/SignificantChangeProjectDatesResponse.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Dfe.PrepareConversions.Data.Models.SignificantChange;
+
+public class SignificantChangeProjectDatesResponse
+{
+	public DateTime? ProposedDecisionDate { get; set; }
+	public DateTime? ProposedChangeDate { get; set; }
+   public SignificantChangeTaskStatus Status { get; set; } = SignificantChangeTaskStatus.NotStarted;
+}