+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Data/Services/SignificantChangeProjectRepository.cs
@@ -164,6 +164,54 @@ public class SignificantChangeProjectRepository(
       }
    }
 
+   public async Task SetReligiousBodyConsultation(int id, SetSignificantChangeReligiousBodyConsultationCommand command)
+   {
+      HttpClient httpClient = httpClientFactory.CreateAcademisationClient();
+      string path = string.Format(PathFor.SetSignificantChangeReligiousBodyConsultation, id);
+
+      var result = await httpClientService.Put<SetSignificantChangeReligiousBodyConsultationCommand, object>(
+         httpClient,
+         path,
+         command);
+
+      if (!result.Success)
+      {
+         throw new ApiResponseException($"Request to Api failed | StatusCode - {result.StatusCode}");
+      }
+   }
+
+   public async Task SetProjectDates(int id, SetSignificantChangeProjectDatesCommand command)
+   {
+      HttpClient httpClient = httpClientFactory.CreateAcademisationClient();
+      string path = string.Format(PathFor.SetSignificantChangeProjectDates, id);
+
+      var result = await httpClientService.Put<SetSignificantChangeProjectDatesCommand, object>(
+         httpClient,
+         path,
+         command);
+
+      if (!result.Success)
+      {
+         throw new ApiResponseException($"Request to Api failed | StatusCode - {result.StatusCode}");
+      }
+   }
+   
+   public async Task SetEqualitiesImpactAssessment(int id, SetSignificantChangeEqualitiesImpactAssessmentCommand command)
+   {
+      HttpClient httpClient = httpClientFactory.CreateAcademisationClient();
+      string path = string.Format(PathFor.SetSignificantChangeEqualitiesImpactAssessment, id);
+
+      var result = await httpClientService.Put<SetSignificantChangeEqualitiesImpactAssessmentCommand, object>(
+         httpClient,
+         path,
+         command);
+
+      if (!result.Success)
+      {
+         throw new ApiResponseException($"Request to Api failed | StatusCode - {result.StatusCode}");
+      }
+   }
+   
    public async Task SetConsultationDuration(int id, SetSignificantChangeConsultationDurationCommand command)
    {
       HttpClient httpClient = httpClientFactory.CreateAcademisationClient();