+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Tests/Pages/SignificantChange/TaskList/SignificantChangeTaskListModelTests.cs
@@ -29,7 +29,8 @@ public class SignificantChangeTaskListModelTests
 
       Assert.IsType<PageResult>(result);
       Assert.NotNull(model.TaskList);
-      Assert.Single(model.TaskList.Sections);
+      Assert.Equal(3, model.TaskList.Sections.Count);
+      Assert.Equal(2, model.TaskList.Sections[0].Tasks.Count);
       repository.Verify(r => r.GetProjectById(id), Times.Once);
    }