+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Utils/SignificantChangeProjectListHelper.cs
@@ -24,7 +24,17 @@ public static class SignificantChangeProjectListHelper
          StakeholderConsultationTrustConsultedStakeholdersNotConsultedReason = significantChangeProject.StakeholderConsultation?.TrustConsultedStakeholdersNotConsultedReason ?? string.Empty,
          ConsultationDurationStatus = significantChangeProject.ConsultationDuration?.Status ?? SignificantChangeTaskStatus.NotStarted,
          ConsultationLastedMinimumThreeWeeks = significantChangeProject.ConsultationDuration?.ConsultationLastedMinimumThreeWeeks,
-         ConsultationDurationNotMetReason = significantChangeProject.ConsultationDuration?.ConsultationDurationNotMetReason ?? string.Empty
+         ConsultationDurationNotMetReason = significantChangeProject.ConsultationDuration?.ConsultationDurationNotMetReason ?? string.Empty,
+         EqualitiesImpactAssessmentStatus = significantChangeProject.EqualitiesImpactAssessment?.Status ?? SignificantChangeTaskStatus.NotStarted,
+         EqualitiesImpactAssessmentCompleted = significantChangeProject.EqualitiesImpactAssessment?.EqualitiesImpactAssessmentCompleted,
+         EqualitiesImpactIdentified = significantChangeProject.EqualitiesImpactAssessment?.EqualitiesImpactIdentified,
+         EqualitiesImpactIdentifiedMitigation = significantChangeProject.EqualitiesImpactAssessment?.EqualitiesImpactIdentifiedMitigation ?? string.Empty,
+         ReligiousBodyConsultationStatus = significantChangeProject.ReligiousBodyConsultation?.Status ?? SignificantChangeTaskStatus.NotStarted,
+         ReligiousBodyConsultationTrustConsultedReligiousBody = significantChangeProject.ReligiousBodyConsultation?.TrustConsultedReligiousBody,
+         ReligiousBodyConsultationTrustConsultedReligiousBodyNotConsultedReason = significantChangeProject.ReligiousBodyConsultation?.TrustConsultedReligiousBodyNotConsultedReason ?? string.Empty,
+         ProjectDatesStatus = significantChangeProject.ProjectDates?.Status ?? SignificantChangeTaskStatus.NotStarted,
+         ProposedDecisionDate = significantChangeProject.ProjectDates?.ProposedDecisionDate,
+         ProposedChangeDate = significantChangeProject.ProjectDates?.ProposedChangeDate
       };
    }