+++ release-notes.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased](https://github.com/DFE-Digital/prepare-academy-conversions/compare/production-2026-08-26.1192...main)
 
 NOTE: remember to update unreleased link when making a new release
-
+### Added
+- [296559](https://dfe-gov-uk.visualstudio.com/Academies-and-Free-Schools-SIP/_boards/board/t/Significant%20changes/Stories?workitem=296559) - Add religious body consultation task for significant change
 
 ---
 ## [1.30.0][1.30.0] - 2026-08-26